+++ src/commands/deleteDisCommand.ts
@@ -1,4 +1,4 @@
-import { Message } from "discord.js";
+import { DMChannel, Message } from "discord.js";
 import Command from "./commandInterface";
 import { Lang } from "./lang/Lang";
 
@@ -9,6 +9,7 @@ export class DeleteDisCommand implements Command {
 	fulldescription: string = this.shortdescription
 	permission: string = ''
 	async run(msg: Message, l: Lang): Promise<void> {
+		if(msg.channel instanceof DMChannel) return
 		await msg.channel.bulkDelete(2)
 	}
